Console sink: Destination enum instead of a bool; build the chunk in Rust
Sink and ConsoleStreamLock take bun_core's Output Destination (Stdout/Stderr)
rather than a to_stderr flag. The stream sink turns its bytes into the Buffer
itself and hands the C++ side a ready chunk, so Bun__NodeWorker__writeConsoleStream
is only "stream.write(chunk), errors ignored".

+++ src/jsc/bindings/ZigGlobalObject.h
@@ -791,7 +791,7 @@ public:
     JSMap* nodeWorkerEnvironmentData() { return m_nodeWorkerEnvironmentData.get(); }
     JSObject* nodeWorkerStdioPorts() { return m_nodeWorkerStdioPorts.get(); }
     void setNodeWorkerStdioPorts(JSObject* ports);
-    JSObject* nodeWorkerConsoleStream(int fd) { return (fd == 2 ? m_nodeWorkerConsoleStderr : m_nodeWorkerConsoleStdout).get(); }
+    JSObject* nodeWorkerConsoleStream(uint8_t fd) { return (fd == 2 ? m_nodeWorkerConsoleStderr : m_nodeWorkerConsoleStdout).get(); }
     void setNodeWorkerConsoleStreams(JSObject* stdoutStream, JSObject* stderrStream);
     void setNodeWorkerEnvironmentData(JSMap* data);
     // node:worker_threads parentPort — the transferred MessagePort entangled with the parent
